+++ s/features/op/loader/parts/ascii-anim.ts
@@ -3,14 +3,12 @@ import {css} from "lit"
 import {nap, repeat} from "@e280/stz"
 
 import {view} from "../../../views/view.js"
+import {Content} from "../../../views/types.js"
 import {cssReset} from "../../../views/css-reset.js"
 
-const style = css`
-:host {
-	font-family: monospace;
-	white-space: pre;
+export function makeAsciiAnim(hz: number, frames: string[]): () => Content {
+	return () => AsciiAnim({hz, frames})
 }
-`
 
 export const AsciiAnim = view(use => ({hz, frames}: {
 		hz: number,
@@ -29,4 +27,12 @@ export const AsciiAnim = view(use => ({hz, frames}: {
 
 	return frames.at(frame())
 })
+
+const style = css`
+:host {
+	font-family: monospace;
+	white-space: pre;
+	user-select: none;
+}
+`