+++ s/features/op/loader/parts/error-display.ts
@@ -3,13 +3,6 @@ import {css, html} from "lit"
 import {view} from "../../../views/view.js"
 import {cssReset} from "../../../views/css-reset.js"
 
-const style = css`
-:host {
-	font-family: monospace;
-	color: red;
-}
-`
-
 export const ErrorDisplay = view(use => (error: any) => {
 	use.name("error")
 	use.styles(cssReset, style)
@@ -23,4 +16,11 @@ export const ErrorDisplay = view(use => (error: any) => {
 	else
 		return `error`
 })
+
+const style = css`
+:host {
+	font-family: monospace;
+	color: red;
+}
+`